Don't run scalatest on init
This caused the tests to be run as desired, but caused clean target and
all other targets dependent on this to fail if ScalaTest run failed.

This was not the desired behaviour.

Still need to find a way to add scalatest as a task before the test task
generates the reports